pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kaggle>=1.7.4.5",
    "kagglehub>=0.4.1",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pyaml>=25.7.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.1",
]
